move tests to associated module
