+++ src/event/format/json.rs
@@ -327,6 +327,8 @@ fn valid_type(data_type: &DataType, value: &Value, schema_version: SchemaVersion
 mod tests {
     use std::str::FromStr;
 
+    use arrow::datatypes::Int64Type;
+    use arrow_array::{ArrayRef, Float64Array, Int64Array, ListArray, StringArray};
     use serde_json::json;
 
     use super::*;
@@ -354,5 +356,451 @@ mod tests {
         let parsed = extract_and_parse_time(&json, "timestamp");
 
         assert!(parsed.is_err());
+    }
+
+    trait TestExt {
+        fn as_int64_arr(&self) -> Option<&Int64Array>;
+        fn as_float64_arr(&self) -> Option<&Float64Array>;
+        fn as_utf8_arr(&self) -> Option<&StringArray>;
+    }
+
+    impl TestExt for ArrayRef {
+        fn as_int64_arr(&self) -> Option<&Int64Array> {
+            self.as_any().downcast_ref()
+        }
+
+        fn as_float64_arr(&self) -> Option<&Float64Array> {
+            self.as_any().downcast_ref()
+        }
+
+        fn as_utf8_arr(&self) -> Option<&StringArray> {
+            self.as_any().downcast_ref()
+        }
+    }
+
+    fn fields_to_map(iter: impl Iterator<Item = Field>) -> HashMap<String, Arc<Field>> {
+        iter.map(|x| (x.name().clone(), Arc::new(x))).collect()
+    }
+
+    #[test]
+    fn basic_object_into_rb() {
+        let json = json!({
+            "c": 4.23,
+            "a": 1,
+            "b": "hello",
+        });
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 1);
+        assert_eq!(rb.num_columns(), 4);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from_iter([1])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from_iter_values(["hello"])
+        );
+        assert_eq!(
+            rb.column_by_name("c").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from_iter([4.23])
+        );
+    }
+
+    #[test]
+    fn basic_object_with_null_into_rb() {
+        let json = json!({
+            "a": 1,
+            "b": "hello",
+            "c": null
+        });
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 1);
+        assert_eq!(rb.num_columns(), 3);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from_iter([1])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from_iter_values(["hello"])
+        );
+    }
+
+    #[test]
+    fn basic_object_derive_schema_into_rb() {
+        let json = json!({
+            "a": 1,
+            "b": "hello",
+        });
+
+        let schema = fields_to_map(
+            [
+                Field::new("a", DataType::Int64, true),
+                Field::new("b", DataType::Utf8, true),
+                Field::new("c", DataType::Float64, true),
+            ]
+            .into_iter(),
+        );
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&schema, false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 1);
+        assert_eq!(rb.num_columns(), 3);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from_iter([1])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from_iter_values(["hello"])
+        );
+    }
+
+    #[test]
+    fn basic_object_schema_mismatch() {
+        let json = json!({
+            "a": 1,
+            "b": 1, // type mismatch
+        });
+
+        let schema = fields_to_map(
+            [
+                Field::new("a", DataType::Int64, true),
+                Field::new("b", DataType::Utf8, true),
+                Field::new("c", DataType::Float64, true),
+            ]
+            .into_iter(),
+        );
+
+        assert!(Event::new(json)
+            .into_recordbatch(&schema, false, None, SchemaVersion::V0,)
+            .is_err());
+    }
+
+    #[test]
+    fn empty_object() {
+        let json = json!({});
+
+        let schema = fields_to_map(
+            [
+                Field::new("a", DataType::Int64, true),
+                Field::new("b", DataType::Utf8, true),
+                Field::new("c", DataType::Float64, true),
+            ]
+            .into_iter(),
+        );
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&schema, false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 1);
+        assert_eq!(rb.num_columns(), 1);
+    }
+
+    #[test]
+    fn array_into_recordbatch_inffered_schema() {
+        let json = json!([
+            {
+                "b": "hello",
+            },
+            {
+                "b": "hello",
+                "a": 1,
+                "c": 1
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c": null
+            },
+        ]);
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 3);
+        assert_eq!(rb.num_columns(), 4);
+
+        let schema = rb.schema();
+        let fields = &schema.fields;
+
+        assert_eq!(&*fields[1], &Field::new("a", DataType::Int64, true));
+        assert_eq!(&*fields[2], &Field::new("b", DataType::Utf8, true));
+        assert_eq!(&*fields[3], &Field::new("c", DataType::Int64, true));
+
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from(vec![None, Some(1), Some(1)])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from(vec![Some("hello"), Some("hello"), Some("hello"),])
+        );
+        assert_eq!(
+            rb.column_by_name("c").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from(vec![None, Some(1), None])
+        );
+    }
+
+    #[test]
+    fn arr_with_null_into_rb() {
+        let json = json!([
+            {
+                "c": null,
+                "b": "hello",
+                "a": null
+            },
+            {
+                "a": 1,
+                "c": 1.22,
+                "b": "hello"
+            },
+            {
+                "b": "hello",
+                "a": 1,
+                "c": null
+            },
+        ]);
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 3);
+        assert_eq!(rb.num_columns(), 4);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from(vec![None, Some(1), Some(1)])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from(vec![Some("hello"), Some("hello"), Some("hello"),])
+        );
+        assert_eq!(
+            rb.column_by_name("c").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from(vec![None, Some(1.22), None,])
+        );
+    }
+
+    #[test]
+    fn arr_with_null_derive_schema_into_rb() {
+        let json = json!([
+            {
+                "c": null,
+                "b": "hello",
+                "a": null
+            },
+            {
+                "a": 1,
+                "c": 1.22,
+                "b": "hello"
+            },
+            {
+                "b": "hello",
+                "a": 1,
+                "c": null
+            },
+        ]);
+
+        let schema = fields_to_map(
+            [
+                Field::new("a", DataType::Int64, true),
+                Field::new("b", DataType::Utf8, true),
+                Field::new("c", DataType::Float64, true),
+            ]
+            .into_iter(),
+        );
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&schema, false, None, SchemaVersion::V0)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 3);
+        assert_eq!(rb.num_columns(), 4);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from(vec![None, Some(1), Some(1)])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from(vec![Some("hello"), Some("hello"), Some("hello"),])
+        );
+        assert_eq!(
+            rb.column_by_name("c").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from(vec![None, Some(1.22), None,])
+        );
+    }
+
+    #[test]
+    fn arr_schema_mismatch() {
+        let json = json!([
+            {
+                "a": null,
+                "b": "hello",
+                "c": 1.24
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c": 1
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c": null
+            },
+        ]);
+
+        let schema = fields_to_map(
+            [
+                Field::new("a", DataType::Int64, true),
+                Field::new("b", DataType::Utf8, true),
+                Field::new("c", DataType::Float64, true),
+            ]
+            .into_iter(),
+        );
+
+        assert!(Event::new(json)
+            .into_recordbatch(&schema, false, None, SchemaVersion::V0,)
+            .is_err());
+    }
+
+    #[test]
+    fn arr_obj_with_nested_type() {
+        let json = json!([
+            {
+                "a": 1,
+                "b": "hello",
+            },
+            {
+                "a": 1,
+                "b": "hello",
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c_a": [1],
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c_a": [1],
+                "c_b": [2],
+            },
+        ]);
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V0)
+            .unwrap();
+        assert_eq!(rb.num_rows(), 4);
+        assert_eq!(rb.num_columns(), 5);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_int64_arr().unwrap(),
+            &Int64Array::from(vec![Some(1), Some(1), Some(1), Some(1)])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from(vec![
+                Some("hello"),
+                Some("hello"),
+                Some("hello"),
+                Some("hello")
+            ])
+        );
+
+        assert_eq!(
+            rb.column_by_name("c_a")
+                .unwrap()
+                .as_any()
+                .downcast_ref::<ListArray>()
+                .unwrap(),
+            &ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
+                None,
+                None,
+                Some(vec![Some(1i64)]),
+                Some(vec![Some(1)])
+            ])
+        );
+
+        assert_eq!(
+            rb.column_by_name("c_b")
+                .unwrap()
+                .as_any()
+                .downcast_ref::<ListArray>()
+                .unwrap(),
+            &ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
+                None,
+                None,
+                None,
+                Some(vec![Some(2i64)])
+            ])
+        );
+    }
+
+    #[test]
+    fn arr_obj_with_nested_type_v1() {
+        let json = json!([
+            {
+                "a": 1,
+                "b": "hello",
+            },
+            {
+                "a": 1,
+                "b": "hello",
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c_a": 1,
+            },
+            {
+                "a": 1,
+                "b": "hello",
+                "c_a": 1,
+                "c_b": 2,
+            },
+        ]);
+
+        let (rb, _) = Event::new(json)
+            .into_recordbatch(&HashMap::default(), false, None, SchemaVersion::V1)
+            .unwrap();
+
+        assert_eq!(rb.num_rows(), 4);
+        assert_eq!(rb.num_columns(), 5);
+        assert_eq!(
+            rb.column_by_name("a").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from(vec![Some(1.0), Some(1.0), Some(1.0), Some(1.0)])
+        );
+        assert_eq!(
+            rb.column_by_name("b").unwrap().as_utf8_arr().unwrap(),
+            &StringArray::from(vec![
+                Some("hello"),
+                Some("hello"),
+                Some("hello"),
+                Some("hello")
+            ])
+        );
+
+        assert_eq!(
+            rb.column_by_name("c_a").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from(vec![None, None, Some(1.0), Some(1.0)])
+        );
+
+        assert_eq!(
+            rb.column_by_name("c_b").unwrap().as_float64_arr().unwrap(),
+            &Float64Array::from(vec![None, None, None, Some(2.0)])
+        );
     }
 }
